refactor(comic-builder): reorganize module structure for clarity
- Extract CharacterCard to dedicated file (was hidden inside ExtractCharactersPhase)
- Create components/phases/ subfolder to separate views from shared components
- Merge renderCharacterThunk into characterThunks for consistent naming
- Standardize projectId access via Redux selector instead of useParams
- Rename index.tsx to ProjectsPage.tsx for clearer intent
