+++ frontend/src/pages/comic-builder/routes.tsx
@@ -1,6 +1,6 @@
 import type { RouteObject } from "react-router";
 import ComicBuilder from "./components/ComicBuilder";
-import Projects from "./index";
+import ProjectsPage from "./ProjectsPage";
 
 // const ROOT_PATH = "/comic-builder";
 const ROOT_PATH = "/";
@@ -8,7 +8,7 @@ const ROOT_PATH = "/";
 export const comicBuilderRoutes: RouteObject[] = [
   {
     path: ROOT_PATH,
-    element: <Projects />,
+    element: <ProjectsPage />,
   },
   {
     path: `${ROOT_PATH}/:projectId`,
