+++ frontend/src/pages/comic-builder/slices/thunks/characterThunks.ts
@@ -18,3 +18,24 @@ export const extractCharacters = createAsyncThunk(
     return response.characters;
   }
 );
+
+type RenderCharacterResponse = {
+  message: string;
+};
+
+type RenderCharacterArgs = {
+  projectId: string;
+  character: Character;
+};
+
+export const renderCharacter = createAsyncThunk(
+  "comicState/renderCharacter",
+  async ({ projectId, character }: RenderCharacterArgs) => {
+    const response = await httpClient.post<RenderCharacterResponse>(
+      ComicBuilderEndpoints.phases.renderCharacter(projectId),
+      character
+    );
+    // No dispatch needed - backend emits state.updated which triggers refetch
+    return response;
+  }
+);
